chore(client): profileをpage.tsxにまとめた
