+++ client/src/app/(authed)/profile/_styles/profile.css.ts
@@ -1,5 +1,15 @@
 import { style } from "@vanilla-extract/css";
 
+export const containerStyle = style({
+  paddingTop: "118px",
+  paddingBottom: "56px",
+});
+
+export const contentStyle = style({
+  margin: "24px 0",
+  padding: "0 24px",
+});
+
 export const headStyle = style({
   display: "flex",
   justifyContent: "space-between",
