--- client/src/app/(authed)/profile/_styles/user-layout.css.ts
@@ -1,11 +1,0 @@
-import { style } from "@vanilla-extract/css";
-
-export const containerStyle = style({
-  paddingTop: "118px",
-  paddingBottom: "56px",
-});
-
-export const contentStyle = style({
-  margin: "24px 0",
-  padding: "0 24px",
-});
